Narrow the incoming coordinate to f32 (#309, #191) (#312)
* Narrow the incoming coordinate to f32 (#309, #191)

The noise machine holds its coordinate values at f32, so an op's x and y are
already f32 when it reads them and the input_scale multiply is an f32
operation on both operands.

Settled by measurement, and by a fixture already committed rather than the
far-field capture that was planned. A DERIVED coordinate leaves the f32 grid
right next to the origin: fulgora_basis is a multioctave read at
wx = ox + wobble_x * wobble_mask, computed in f64 and off the f32 grid at 55
of that fixture's 101 positions. Scoring sumOctaves with and without its
incoming narrowing:

  narrowed    (shipped)   101/101, worst residual exactly 0
  un-narrowed (planted)    81/101, worst 7.0333e-6

Twenty positions discriminate. That is also the check #191 asked for in its
own words - "a caller that passes a derived coordinate" - which Fulgora has
satisfied since it landed.

What changed:

- basisNoiseExpr narrows x and y before the input_scale multiply, which is
  what the Rust already did. #309 was the two ports disagreeing about this;
  they now agree. offset_x is 0 at all six call sites, so whether the game
  narrows before or after that add is unreachable and no third variant is
  invented for it.
- variablePersistenceMultioctaveNoise narrows y, in BOTH ports. It narrowed x
  only as a side effect of the f32(x + offset_x) add; y had no add and so was
  silently multiplied in f64. Both ports had the identical asymmetry, which is
  why tier 2 could not see it - they agreed with each other while both
  disagreed with the game.
- quickMultioctaveNoise needed nothing. #191's issue text is stale: it already
  narrows both coordinates at quickMultioctaveNoise.ts:192-193.
- basisNoise itself is deliberately unchanged. Its remaining direct callers
  (nauvisShared.ts:133, startingPatches.ts:185, regularPatches.ts:164) are
  unported Nauvis chains where the whole expression is un-narrowed, not just
  the coordinate. Planting the internal narrowing leaves all 26 of their specs
  passing, so no committed fixture discriminates it. That belongs to #226,
  scored layer by layer under the greedy-accept rule.

The tier-2 pin is INVERTED rather than deleted. "the two ports agree off the
f32 grid" now asserts 0 of 74 diverging where it froze 32, and reverting
primitives.ts reproduces [ 'hairlineCracks', ...(31) ] - planted, not assumed.
Its anti-vacuity is not optional: "nothing diverges" is exactly what a sweep
evaluating nothing would report, so it also asserts the off-grid window's
folds differ from the on-grid window's on all 74 fields.

Two lessons recorded in CLAUDE.md, the second bigger than the bug: fold the
fixtures you already have before capturing more, and "no fixture can grade
this" is a claim about the fixtures you looked at. It was true of every
Vulcanus fixture and false of a Fulgora one.

pnpm run verify green: 2143 passed, 4 skipped, verify:rust included.


* Act on the three tier-2 review points from #310

All three were flagged when Vulcanus tier 2 landed and left for a follow-up.

Move the field selector into fmw-noise. VulcanusParity now sits beside
VulcanusStack in expressions/vulcanus_stack.rs, and the wasm export builds the
stack through the render helpers and calls parity.field(field, x, y). The
reason is ownership of test-only API: the selector needs elevation_fields and
temperature, which no render path reads, and reaching them from another crate
meant two pub methods on a library type existing solely for a test - and a pub
method cannot be #[cfg(test)]-gated because the wasm crate calls it at build
time. Both are private again. VULCANUS_FIELD_COUNT moved with it as
VulcanusParity::FIELD_COUNT, so the count and the match it bounds cannot drift.

This is pure code motion and was checked as such rather than assumed: tier 2
(74 fields, two sliders, two windows) and tier 3 (byte-identical renders) both
pass unchanged, 28 tests. It does move engine.wasm by 142 bytes, because the
selector inlines differently once it is in the same crate as the layers it
reads - a reminder that a wasm diff is not by itself evidence of a behaviour
change. tile_index moved verbatim, NAN fallback included; writing -1.0 there
would have been a behaviour change smuggled into a code move.

Freeze the slider anti-vacuity guard at its measured 50 of 74 instead of
leaving it as "> 40". The argument for a floor was that the count is a property
of which fields read a slider rather than a result - but that is equally true
of the window guard's 74 and the tile guard's 19, and both are frozen. The
freeze is strictly stronger: a floor of 40 would sit through nine fields
silently dropping their slider dependency.

Record verify:rust's cost as a RANGE, 1m44s to 2m48s. #310 measured 1m44s on CI
against the 2m48s recorded in CLAUDE.md, on code whose Rust half did not change
between the two runs. That is the same spread the test shards show, so the
honest response is to widen the figure rather than replace it with whichever
run was seen last.

pnpm run verify green: 2143 passed, 4 skipped, verify:rust included.


---------

+++ src/noise/eval/primitives.ts
@@ -50,10 +50,36 @@ export interface BasisExprParams {
  * That is why the older `oracle-basis` fixture, captured at `output_scale = 1`,
  * could not answer the question and a new capture had to.
  *
- * The `input_scale` product is deliberately NOT narrowed here. It decides which
- * point gets sampled rather than what the product rounds to, and narrowing it
- * would evaluate a DIFFERENT point. Whether the game holds `input_scale` at f32
- * too is a separate, unmeasured question tracked on #269.
+ * ## The incoming coordinate is narrowed too (#309, #191)
+ *
+ * The noise machine holds its coordinate VALUES at f32, so `x` and `y` are
+ * already f32 by the time this op reads them and the `input_scale` multiply is
+ * an f32 operation on both operands. That was settled by measurement rather
+ * than by argument, and by an EXISTING fixture rather than a new capture:
+ * `fulgora_basis` is a multioctave evaluated at Fulgora's distorted coordinate
+ * `wx = ox + wobble_x * wobble_mask`, which leaves the f32 grid at 55 of that
+ * fixture's 101 positions. Scored against the game:
+ *
+ * ```text
+ * narrowed    (f32(x) before the scale)   101/101, worst residual exactly 0
+ * un-narrowed (f64 x)                      81/101, worst 7.0333e-6
+ * ```
+ *
+ * 20 positions discriminate, so this is a measurement and not a preference. It
+ * is also exactly the check #191 asked for in its own words - a caller passing
+ * a DERIVED coordinate - which Fulgora has satisfied since it landed, unnoticed.
+ *
+ * **No fixture can grade this op's narrowing DIRECTLY, and that is the trap.**
+ * The game snaps every sample onto its 1/256 MapPosition grid (#186), and below
+ * |x| = 65536 that grid is a SUBSET of the f32 grid - so at every position a
+ * fixture can offer, the narrowed and un-narrowed forms are bit-identical.
+ * Scoring at the RAW, unsnapped fixture coordinates looks like a discriminating
+ * measurement and is not: those are points the game never evaluated. Snap with
+ * `test/captureGrid.ts` first. The evidence has to come from a caller that
+ * derives its coordinate, which is why it came from Fulgora.
+ *
+ * `offset_x` is 0 at all six call sites, so whether the game narrows before or
+ * after that add is unreachable, and no third variant is invented for it.
  */
 export function basisNoiseExpr(
   x: number,
@@ -65,6 +91,6 @@ export function basisNoiseExpr(
   const inputScale = f32(params.inputScale);
   return f32(
     f32(params.outputScale) *
-      basisNoise(f32((x + offsetX) * inputScale), f32(y * inputScale), tables),
+      basisNoise(f32(f32(x + offsetX) * inputScale), f32(f32(y) * inputScale), tables),
   );
 }
